refactor (path): renommer BeginPath en BeginTraining (#59)
