--- src/Path/BeginPath/PathAlreadyBeganException.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace IncentiveFactory\Game\Path\BeginPath;
-
-use InvalidArgumentException;
-
-final class PathAlreadyBeganException extends InvalidArgumentException
-{
-}
