+++ src/Path/BeginTraining/BeginTraining.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace IncentiveFactory\Game\Path\BeginPath;
+namespace IncentiveFactory\Game\Path\BeginTraining;
 
 use DateTimeImmutable;
 use IncentiveFactory\Game\Path\Path;
@@ -10,16 +10,16 @@ use IncentiveFactory\Game\Path\PathGateway;
 use IncentiveFactory\Game\Shared\Command\CommandHandler;
 use IncentiveFactory\Game\Shared\Uid\UlidGeneratorInterface;
 
-final class BeginPath implements CommandHandler
+final class BeginTraining implements CommandHandler
 {
     public function __construct(private PathGateway $pathGateway, private UlidGeneratorInterface $ulidGenerator)
     {
     }
 
-    public function __invoke(BeginningOfPath $beginningOfPath): void
+    public function __invoke(BeginningOfTraining $beginningOfPath): void
     {
         if ($this->pathGateway->hasAlreadyBegan($beginningOfPath->player, $beginningOfPath->training)) {
-            throw new PathAlreadyBeganException('Path already began');
+            throw new TrainingAlreadyBeganException('Path already began');
         }
 
         $path = Path::create(
