+++ src/Path/BeginTraining/TrainingAlreadyBeganException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace IncentiveFactory\Game\Path\BeginTraining;
+
+use InvalidArgumentException;
+
+final class TrainingAlreadyBeganException extends InvalidArgumentException
+{
+}
